requirements.txt
# =====================================================
# ARIS - Python Dependencies
# =====================================================

# Core data handling
# Used for reading, cleaning, and aggregating UIDAI data
pandas>=1.5.0
numpy>=1.23.0

# API handling
# Used to call data.gov.in APIs
requests>=2.31.0

# Environment variable support
# Used to safely load API keys from .env
python-dotenv>=1.0.0

# Frontend dashboard
# Used to build interactive UI
streamlit>=1.30.0

# Visualisation
# Used to draw India map and heatmaps
plotly>=5.0.0
